pyproject.toml
[project]
name = "neo4j-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "mcp[cli]>=1.18.0",
    "neo4j>=6.0.2",
    "openai>=2.6.0",
    "pydantic-settings>=2.11.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
]
